refactor: UI functions separated in files for better organization
